requirements.txt
fastapi 
uvicorn 
mlflow 
transformers 
torch
motor
pymongo
passlib[bcrypt]
python-jose
